hw11

+++ core-service/src/main/java/com/geekbrains/spring/web/entities/Product.java
@@ -7,6 +7,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -25,7 +26,7 @@ public class Product {
     private String title;
 
     @Column(name = "price")
-    private Integer price;
+    private BigDecimal price;
 
     @ManyToOne(targetEntity = Category.class)
     @JoinColumn(name = "category_id")
@@ -42,7 +43,7 @@ public class Product {
     @OneToMany(mappedBy = "product")
     private List<OrderItem> orderItems;
 
-    public Product(Long id, String title, Integer price) {
+    public Product(Long id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
